Fix back button position; make group for screens in multiple navigators; closes #808
