+++ src/components/MediaViewer/MediaViewer.js
@@ -1,9 +1,8 @@
 // @flow
 
-import { useFocusEffect, useNavigation } from "@react-navigation/native";
+import { useNavigation } from "@react-navigation/native";
 import classnames from "classnames";
 import { Heading4, TransparentCircleButton, WarningSheet } from "components/SharedComponents";
-import BackButton from "components/SharedComponents/Buttons/BackButton";
 import { SafeAreaView, View } from "components/styledComponents";
 import { ObsEditContext } from "providers/contexts";
 import type { Node } from "react";
@@ -11,7 +10,7 @@ import React, {
   useCallback, useContext, useEffect, useRef,
   useState
 } from "react";
-import { BackHandler, StatusBar } from "react-native";
+import { StatusBar } from "react-native";
 import { BREAKPOINTS } from "sharedHelpers/breakpoint";
 import useDeviceOrientation from "sharedHooks/useDeviceOrientation";
 import useTranslation from "sharedHooks/useTranslation";
@@ -54,32 +53,6 @@ const MediaViewer = ( ): Node => {
   const showWarningSheet = ( ) => setWarningSheet( true );
   const hideWarningSheet = ( ) => setWarningSheet( false );
 
-  const handleBackButtonPress = useCallback( ( ) => navigation.goBack( ), [navigation] );
-
-  const renderBackButton = useCallback( ( ) => (
-    <View className="ml-4">
-      <BackButton
-        color={colors.white}
-        onPress={handleBackButtonPress}
-      />
-    </View>
-  ), [handleBackButtonPress] );
-
-  useFocusEffect(
-    useCallback( ( ) => {
-      // make sure an Android user cannot back out to MyObservations with the back arrow
-      // and see a stale observation context state
-      const onBackPress = ( ) => {
-        handleBackButtonPress( );
-        return true;
-      };
-
-      BackHandler.addEventListener( "hardwareBackPress", onBackPress );
-
-      return ( ) => BackHandler.removeEventListener( "hardwareBackPress", onBackPress );
-    }, [handleBackButtonPress] )
-  );
-
   const deletePhoto = ( ) => {
     deletePhotoFromObservation( mediaViewerUris[selectedPhotoIndex] );
     hideWarningSheet( );
@@ -96,12 +69,11 @@ const MediaViewer = ( ): Node => {
     );
 
     const headerOptions = {
-      headerTitle: renderHeaderTitle,
-      headerLeft: renderBackButton
+      headerTitle: renderHeaderTitle
     };
 
     navigation.setOptions( headerOptions );
-  }, [navigation, t, numOfPhotos, renderBackButton] );
+  }, [navigation, t, numOfPhotos] );
 
   const scrollToIndex = useCallback( index => {
     // when a user taps a photo in the carousel, the UI needs to automatically
